Fix some specimen caching mismatches
